requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
mysqlclient==2.2.4
PyJWT==2.9.0
sqlparse==0.5.0
tzdata==2024.1
